backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cssselect2==0.7.0
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-rest-swagger==2.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
fonttools==4.55.2
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
Markdown==3.7
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zopfli==0.2.3.post1
gunicorn

